refactor(web/applications): move AD methods to a separate service module
