+++ apps/web/src/server/applications/case/project-team/applications-project-team.controller.js
@@ -1,7 +1,5 @@
-import {
-	getManyProjectTeamMembersInfo,
-	getAllCachedUsers
-} from '../../common/services/project-team.service.js';
+import { getManyProjectTeamMembersInfo } from '../../common/services/project-team.service.js';
+import projectADService from '../../common/services/azure-project-team.js';
 import {
 	destroySuccessBanner,
 	getSuccessBanner,
@@ -198,7 +196,7 @@ export async function viewProjectTeamSearchPage(
 	}
 
 	// checkpoint 2: retrieve all azure users from cache or return 401
-	const allAzureUsers = await getAllCachedUsers(session);
+	const allAzureUsers = await projectADService.getAllCachedUsers(session);
 
 	// checkpoint 3: search member and render paginated list of results
 	const searchResults = await searchProjectTeamMembersData(
